# hardware/server/openapi.yaml
...
openapi: 3.1.0
info:
  title: SitSense API (mock)
  description: 'Raspberry Pi 서버 → 안드로이드 앱. 같은 Wi-Fi에서 `http://<Pi IP>:8000`.


    지금은 **mock 서버**: 시나리오 더미 데이터와 가짜 판단 결과를 준다. 시각은 time.time() 초, 거리 mm, 각도 °, 비율
    0~1. 표시용 글자는 앱이 만든다. 영상 이미지는 어떤 응답에도 없다.'
  version: 0.1.0-mock
paths:
...
  /calibrate:
    post:
      tags:
      - 기준 자세
      summary: 기준 자세 측정 (앉으면 10초). 세션 없이(온보딩) 또는 세션 중(설정의 다시 측정)
      operationId: calibrate_calibrate_post
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CalibrateResult'
        '409':
          description: 지금 상태에서 할 수 없음 (detail.code 참고)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: 잘못된 값 (detail.code 참고)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /time/sync:
    post:
      tags:
      - 시각
      summary: 앱 시각·시간대 보내기 (연결할 때, 세션 시작 전마다)
      operationId: time_sync_time_sync_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TimeSync'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TimeStatus'
        '409':
          description: 지금 상태에서 할 수 없음 (detail.code 참고)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: 잘못된 값 (detail.code 참고)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /session/start:
    post:
      tags:
      - 세션
      summary: 측정 시작 (시각 동기화 후에만). 기준 자세 측정부터
      operationId: session_start_session_start_post
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionStatus'
        '409':
          description: 지금 상태에서 할 수 없음 (detail.code 참고)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: 잘못된 값 (detail.code 참고)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /session/pause:
    post:
      tags:
      - 세션
      summary: '일시정지: 기록·판단 멈춤, 카메라·거리 끔 (착석 구간은 여기서 끝)'
      operationId: session_pause_session_pause_post
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionStatus'
        '409':
          description: 지금 상태에서 할 수 없음 (detail.code 참고)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: 잘못된 값 (detail.code 참고)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /session/resume:
    post:
      tags:
      - 세션
      summary: 재개
      operationId: session_resume_session_resume_post
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionStatus'
        '409':
          description: 지금 상태에서 할 수 없음 (detail.code 참고)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: 잘못된 값 (detail.code 참고)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /session/stop:
    post:
      tags:
      - 세션
      summary: '정지: 세션 끝, 센서 모두 끔 (앉아도 자동으로 다시 시작하지 않음)'
      operationId: session_stop_session_stop_post
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionStatus'
        '409':
          description: 지금 상태에서 할 수 없음 (detail.code 참고)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: 잘못된 값 (detail.code 참고)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /records/reset:
    post:
      tags:
      - 기록
      summary: 기록 초기화 (설정·기준 자세는 유지). 세션이 없을 때만, 확인 문구 필요
      operationId: records_reset_records_reset_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RecordsReset'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecordsResetResult'
        '409':
          description: 지금 상태에서 할 수 없음 (detail.code 참고)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: 잘못된 값 (detail.code 참고)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
...
    BaselinePhase:
      type: string
      enum:
      - none
      - waiting_seat
      - measuring
      - ready
      - failed
      title: BaselinePhase
...
    CalibrateResult:
      properties:
        started:
          type: boolean
          title: Started
        seconds:
          type: number
          title: Seconds
          description: 측정에 걸리는 초 (이 동안 바른 자세로 앉기)
      type: object
      required:
      - started
      - seconds
      title: CalibrateResult
...
    ErrorDetail:
      properties:
        code:
          type: string
          title: Code
          description: 앱이 처리할 오류 코드 (CLOCK_NOT_SYNCED, CLOCK_BACKWARD, CLOCK_JUMP_IN_SESSION,
            CLOCK_BEFORE_LAST_RECORD, SESSION_ACTIVE, NOT_RUNNING, NOT_PAUSED, NO_SESSION,
            CONFIRM_REQUIRED, INVALID_TIME, INVALID_TIMEZONE)
        message:
          type: string
          title: Message
          description: 개발자용 설명 (앱 화면 문구는 앱이 만든다)
        resync_required:
          type: boolean
          title: Resync Required
          description: true면 POST /time/sync를 다시 보내야 함
        boot_id:
          type: string
          title: Boot Id
          description: 서버가 켜질 때마다 바뀌는 값. 앱이 기억한 값과 다르면 재시작된 것
        last_record_at:
          anyOf:
          - type: number
          - type: 'null'
          title: Last Record At
          description: CLOCK_BEFORE_LAST_RECORD일 때 Pi의 마지막 기록 시각 (앱이 날짜를 보여줄 때)
      type: object
      required:
      - code
      - message
      - resync_required
      - boot_id
      title: ErrorDetail
    ErrorResponse:
      properties:
        detail:
          $ref: '#/components/schemas/ErrorDetail'
      type: object
      required:
      - detail
      title: ErrorResponse
      description: 오류 응답 (409 상태 충돌, 422 잘못된 값)
...
    RecordsReset:
      properties:
        confirm:
          type: string
          title: Confirm
          description: 반드시 'DELETE_RECORDS'
          examples:
          - DELETE_RECORDS
      type: object
      required:
      - confirm
      title: RecordsReset
      description: 기록 초기화 요청 — 앱에서 확인 절차를 거친 뒤 보낸다
    RecordsResetResult:
      properties:
        deleted:
          additionalProperties:
            type: integer
          type: object
          title: Deleted
          description: 테이블별 지운 줄 수 (설정·기준 자세는 유지)
        resync_required:
          type: boolean
          title: Resync Required
          description: true — 시각 동기화를 지웠으니 POST /time/sync를 다시 보내야 함
      type: object
      required:
      - deleted
      - resync_required
      title: RecordsResetResult
...
    SensorPower:
      type: string
      enum:
      - 'on'
      - 'off'
      - warming
      title: SensorPower
...
    SessionBaseline:
      properties:
        state:
          $ref: '#/components/schemas/BaselinePhase'
          description: none 기준 없음 / waiting_seat 앉기를 기다림 / measuring 측정 중 / ready
            기준 있음 / failed 마지막 측정 실패
        kind:
          anyOf:
          - type: string
          - type: 'null'
          title: Kind
          description: 측정 중이거나 마지막 측정의 종류 (initial / session / recalibration)
        remaining_sec:
          anyOf:
          - type: number
          - type: 'null'
          title: Remaining Sec
          description: 측정 중이면 남은 초 (앉기를 기다리는 중이면 측정 시간 전체)
        fail_reason:
          anyOf:
          - type: string
          - type: 'null'
          title: Fail Reason
          description: 실패 이유 (no_person / sensor_lost / too_much_motion)
        using_previous:
          type: boolean
          title: Using Previous
          description: 이번 측정이 실패해서 직전 성공 기준으로 판단 중인지
        measured_at:
          anyOf:
          - type: number
          - type: 'null'
          title: Measured At
          description: 지금 쓰는 기준을 잰 시각
      type: object
      required:
      - state
      - using_previous
      title: SessionBaseline
      description: 세션의 기준 자세 — 시작 버튼을 누르면 앉은 게 확인된 뒤 10초 측정 (재개 때는 다시 재지 않음)
    SessionSensors:
      properties:
        pressure:
          $ref: '#/components/schemas/SensorPower'
        camera:
          $ref: '#/components/schemas/SensorPower'
          description: warming = 켰지만 아직 사람을 찾기 전 (약 2초)
        distance:
          $ref: '#/components/schemas/SensorPower'
      type: object
      required:
      - pressure
      - camera
      - distance
      title: SessionSensors
      description: 센서 전원 (자리 비움 5초면 카메라·거리 끔, 일시정지면 압력만, 정지면 모두 끔)
    SessionState:
      type: string
      enum:
      - idle
      - running
      - paused
      title: SessionState
    SessionStatus:
      properties:
        state:
          $ref: '#/components/schemas/SessionState'
          description: idle 세션 없음 / running 측정 중 / paused 일시정지
        session_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Session Id
          description: 마지막(또는 지금) 세션 ID. 한 번도 안 했으면 null
        started_at:
          anyOf:
          - type: number
          - type: 'null'
          title: Started At
        ended_at:
          anyOf:
          - type: number
          - type: 'null'
          title: Ended At
          description: 정지한 시각. 진행 중이면 null
        elapsed_sec:
          anyOf:
          - type: number
          - type: 'null'
          title: Elapsed Sec
          description: 경과 시간 (일시정지 시간 제외)
        paused_sec:
          anyOf:
          - type: number
          - type: 'null'
          title: Paused Sec
          description: 일시정지한 시간 합
        seated_sec:
          anyOf:
          - type: number
          - type: 'null'
          title: Seated Sec
          description: 이 세션의 착석 시간 (착석 구간 합, 일시정지 제외)
        seated_now:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Seated Now
          description: 지금 압력이 있는지 (일시정지 중에도). 압력센서가 꺼져 있으면 null
        sensors:
          $ref: '#/components/schemas/SessionSensors'
        clock_synced:
          type: boolean
          title: Clock Synced
        boot_id:
          type: string
          title: Boot Id
        end_reason:
          anyOf:
          - type: string
          - type: 'null'
          title: End Reason
          description: 마지막 세션이 끝난 이유 (stop / baseline_failed). 진행 중이면 null
        baseline:
          $ref: '#/components/schemas/SessionBaseline'
      type: object
      required:
      - state
      - sensors
      - clock_synced
      - boot_id
      - baseline
      title: SessionStatus
      description: 측정 세션 상태 — 홈 하단 측정 바
...
    TimeStatus:
      properties:
        synced:
          type: boolean
          title: Synced
          description: 이 서버가 켜진 뒤 동기화했는지
        boot_id:
          type: string
          title: Boot Id
          description: 서버가 켜질 때마다(재부팅 포함) 바뀜. 바뀌었으면 다시 동기화
        timezone:
          anyOf:
          - type: string
          - type: 'null'
          title: Timezone
        app_time:
          anyOf:
          - type: number
          - type: 'null'
          title: App Time
          description: 앱 기준 지금 시각. 동기화 전이면 null
        local_time:
          anyOf:
          - type: string
          - type: 'null'
          title: Local Time
          description: app_time을 시간대로 바꾼 ISO 문자열 (확인용)
        synced_at:
          anyOf:
          - type: number
          - type: 'null'
          title: Synced At
          description: 마지막 동기화 때 앱이 보낸 시각
        resync_required:
          type: boolean
          title: Resync Required
          description: true면 POST /time/sync 필요
        offset_change_sec:
          anyOf:
          - type: number
          - type: 'null'
          title: Offset Change Sec
          description: 다시 동기화했을 때 시각이 바뀐 양 (초). 첫 동기화·조회에서는 null
        applied:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Applied
          description: 이번 동기화를 적용했는지. 0.5초 이내로 뒤로 가는 변경은 false(시각 그대로). 조회에서는 null
        message:
          anyOf:
          - type: string
          - type: 'null'
          title: Message
      type: object
      required:
      - synced
      - boot_id
      - resync_required
      title: TimeStatus
      description: Pi의 시각 동기화 상태. Pi OS 시계는 바꾸지 않고 '앱 시각 − Pi 단조시계' 차이만 저장
    TimeSync:
      properties:
        app_time:
          type: number
          title: App Time
          description: 앱의 현재 시각 (time.time()과 같은 초, 소수 가능)
          examples:
          - 1791207600.123
        timezone:
          type: string
          title: Timezone
          description: 시간대 (Asia/Seoul만 가능)
          default: Asia/Seoul
          examples:
          - Asia/Seoul
      type: object
      required:
      - app_time
      title: TimeSync
      description: '앱 → Pi: 현재 시각과 시간대. 연결할 때와 세션 시작 전마다 보낸다'
...